lunaris-orion/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0
pytorch-lightning>=2.1.0
datasets>=2.15.0
pillow>=10.1.0
numpy>=1.24.0
fastapi>=0.105.0
uvicorn>=0.24.0
python-jose>=3.3.0
python-multipart>=0.0.6
discord.py>=2.3.0
stripe>=7.0.0
prometheus-client>=0.19.0
wandb>=0.16.0
opencv-python>=4.8.0
albumentations>=1.3.1
pandas>=2.1.0
pyarrow>=14.0.1
requests>=2.31.0
tqdm>=4.66.1 